src/main/java/pipe/actions/edit/RedoAction.java
package pipe.actions.edit;

import pipe.actions.GuiAction;
import pipe.controllers.PetriNetController;
import pipe.controllers.PipeApplicationController;
import pipe.gui.ApplicationSettings;

import java.awt.event.ActionEvent;

public class RedoAction extends GuiAction {

    public RedoAction() {
        super("Redo", "Redo (Ctrl-Y)", "ctrl Y");
    }

    @Override
    public void actionPerformed(final ActionEvent actionEvent) {
        PipeApplicationController applicationController = ApplicationSettings
                .getApplicationController();
        PetriNetController controller = applicationController.getActivePetriNetController();
        controller.getHistoryManager().doRedo();
    }
}